src/main/java/com/turnero/turnos/entity/DomicilioDTO.java
package com.turnero.turnos.entity;

public class DomicilioDTO {
    private Long id;
    private String calle;
    private Integer numero;
    private String ciudad;
    private String provincia;

    public DomicilioDTO( String calle, Integer numero, String ciudad, String provincia) {

        this.calle = calle;
        this.numero = numero;
        this.ciudad = ciudad;
        this.provincia = provincia;
    }
    public DomicilioDTO(){

    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getCalle() {
        return calle;
    }

    public void setCalle(String calle) {
        this.calle = calle;
    }

    public Integer getNumero() {
        return numero;
    }

    public void setNumero(Integer numero) {
        this.numero = numero;
    }

    public String getCiudad() {
        return ciudad;
    }

    public void setCiudad(String ciudad) {
        this.ciudad = ciudad;
    }

    public String getProvincia() {
        return provincia;
    }

    public void setProvincia(String provincia) {
        this.provincia = provincia;
    }
}
